[automation] Move trainCombat function to a new module
